+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+# Head
+
+-   Added: supported `manifest` plugin option for export source and interpolated filenames.
+
 # 1.0.1 - 2016-03-15
 
 -   Fixed: don't duplicate excluded assets, less memory usage.